requirements.txt
Flask
Flask-Cors
python-dotenv
openai==0.28.0
gunicorn
